Remove a cycle dependency
