+++ src/main/resources/application.yaml
@@ -12,7 +12,7 @@ spring:
   datasource:
     driver-class-name: org.postgresql.Driver
     url: jdbc:postgresql://${BULK_SCANNING_DB_HOST:localhost}:${BULK_SCANNING_DB_PORT:5435}/${BULK_SCANNING_DB_NAME:bulkscans}${BULK_SCANNING_DB_CONN_OPTIONS:}
-    username: ${BULK_SCANNING_DB_USER_NAME:bulkscans}
+    username: ${BULK_SCANNING_DB_USER_NAME:bulkscan}
     password: ${BULK_SCANNING_DB_PASSWORD:}
     properties:
       charSet: UTF-8
